Fix Cancel designator not activated when the thing has a designation.
When we add destruct command to a wall, we restart the cycle every time and the build copy command is in the first of the list. As a result, pressing key will always return build copy command of the wall, not the cancel command. We place the latter in the first place in this case.